add tree cover to aoi dashboards
