--- app/javascript/components/widgets/land-cover/tree-cover/settings.js
@@ -1,4 +1,0 @@
-export default {
-  threshold: 30,
-  extentYear: 2000
-};
